Fix: renamed component package to components
